Add functions for calculating and plotting gravitational potential
The gravity_potential function takes the mass of the largest object
m1, the relative mass m2, its orbital radius, and the x and y
coordinates of the grid being plotted, and calculates the gravitational
potential at each grid point. The gravity_potential function is currently
being called by an if __name__=="__main__" block which allows any portable
code to be tested when running this program while allowing the other
functions defined to be called without running the tests when this
file is imported as a python module.