Add placeholder image preview of project
Stock image of a gravity well as a placeholder project preview. Since the
Markdown code used to add this image to the displayed README.md file is
dependent on the PATH, if the image does not display properly on the
remote repository, the Markdown code will need to be further refined. As a
stopgap measure, the image can be added to the README.md file on the remote
repository and committed directly, although this is less than ideal.