Add blank Qt app
The app consists of a QMainWindow only right now, but this will serve as
the "backbone" of the app as it is being developed. The various GUI
features will be added to pages on a QStackedWidget and the various
functions used in the simulation programs will be imported and added to
methods of the QMainWindow, allowing the user to plot the gravitational
potential and trajectory on a matplotlib added as a widget to the GUI.
Buttons and input only text fields will also be added so that users
can specify parameters of the simulation.