Fixed gravitational potential generator
The original script used for calculating the gravitational potential at
each point on a set grid was intended for a specific ratio of the two large
masses. The script has been fixed so that it works properly with other
ratios. Since it is expected that the gravity_potential function will
receive a wide range of values for its parameters, it will now throw
exceptions for parameters that don't make physical sense being passed.